dragon_dictation_pro/requirements.txt

# backend/requirements.txt
ctranslate2
flask
waitress
faster-whisper
google-generativeai
requests==2.31.0